+++ srcjs/__tests__/Reactable.test.js
@@ -9621,7 +9621,7 @@ describe('reactable JavaScript API', () => {
     expect(queryAllByText('detail')).toHaveLength(0)
   })
 
-  it('Reactable.downloadDataCSV', () => {
+  it('Reactable.downloadDataCSV and Reactable.getDataCSV', () => {
     const props = {
       data: { a: ['a11', 'a12', 'a23'], b: [2, 3, 3] },
       columns: [
@@ -9641,12 +9641,23 @@ describe('reactable JavaScript API', () => {
     expect(downloadCSV).toHaveBeenCalledTimes(1)
     expect(downloadCSV).toHaveBeenLastCalledWith('a,b\na11,2\na12,3\na23,3\n', 'data.csv')
 
+    // Custom filename
     reactable.downloadDataCSV('my-tbl', 'my_custom_filename.csv')
     expect(downloadCSV).toHaveBeenCalledTimes(2)
     expect(downloadCSV).toHaveBeenLastCalledWith(
       'a,b\na11,2\na12,3\na23,3\n',
       'my_custom_filename.csv'
     )
+
+    // Custom options
+    const options = { columnIds: ['b', 'a'], headers: false, sep: '\t' }
+    reactable.downloadDataCSV('my-tbl', null, options)
+    expect(downloadCSV).toHaveBeenCalledTimes(3)
+    expect(downloadCSV).toHaveBeenLastCalledWith('2\ta11\n3\ta12\n3\ta23\n', 'data.csv')
+
+    // getDataCSV
+    const csv = reactable.getDataCSV('my-tbl', options)
+    expect(csv).toEqual('2\ta11\n3\ta12\n3\ta23\n')
 
     // Should download filtered data
     const searchInput = getSearchInput(container)
@@ -9655,13 +9666,13 @@ describe('reactable JavaScript API', () => {
     const sortableHeaders = getSortableHeaders(container)
     fireEvent.click(sortableHeaders[0])
     reactable.downloadDataCSV('my-tbl')
-    expect(downloadCSV).toHaveBeenCalledTimes(3)
+    expect(downloadCSV).toHaveBeenCalledTimes(4)
     expect(downloadCSV).toHaveBeenLastCalledWith('a,b\na11,2\na12,3\n', 'data.csv')
 
     // Should use flattened rows and exclude aggregated rows when grouped
     rerender(<Reactable {...props} groupBy={['b']} />)
     reactable.downloadDataCSV('my-tbl')
-    expect(downloadCSV).toHaveBeenCalledTimes(4)
+    expect(downloadCSV).toHaveBeenCalledTimes(5)
     expect(downloadCSV).toHaveBeenLastCalledWith('a,b\na11,2\na12,3\n', 'data.csv')
   })